+++ gulpfile.js
@@ -4,7 +4,6 @@ const del = require('del');
 const ts = require('gulp-typescript');
 const mocha = require('gulp-spawn-mocha');
 const sourcemaps = require('gulp-sourcemaps');
-const remapIstanbul = require('remap-istanbul/lib/gulpRemapIstanbul');
 const merge = require('merge2');
 
 const scripts = ['src/**/*.ts'];
@@ -44,27 +43,14 @@ const tests = () => gulp.src(['dist/test/**/*.js'], { read: false })
 		this.emit('end');
 	});
 
-const coverage = () => gulp.src(['dist/test/**/*.js'], { read: false })
-	.pipe(mocha({
-		reporter: 'dot',
-		timeout: 10000,
-		istanbul: { print: 'none' },
-	}));
-
-const useCoverage = process.argv.indexOf('--coverage') !== -1;
-
 const watch = cb => {
 	gulp.watch([
 		...scripts, 'test/**/*.psd', 'test/**/*.psb', 'test/**/*.abr', 'test/**/*.png', 'test/**/*.json'
-	], useCoverage ? cov : test);
+	],  test);
 	cb();
 };
 
-const remap = () => gulp.src('coverage/coverage.json')
-	.pipe(remapIstanbul({ reports: { html: 'coverage-remapped' } }));
-
 const dev = gulp.series(clean, buildJS, watch);
-const cov = gulp.series(buildJS, coverage, remap);
 const test = gulp.series(buildJS, tests);
 
-module.exports = { build, buildJS, buildES, dev, cov, test };
+module.exports = { build, buildJS, buildES, dev, test };