+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v14.5.0
+- Added support for zip compression without prediction of image data
+- Added support for `compress` option when writing
+
 ## v14.4.0
 - Added support for `layerMaskAsGlobalMask` flag
 - Added support for `interpolationMethod` for gradient overlay layer blending mode